Remove dependency of re_log_encoding on re_log_channel
