+++ crates/store/re_data_source/src/lib.rs
@@ -7,6 +7,7 @@
 //! Also handles different file types: rrd, images, text files, 3D models, point clouds…
 
 mod data_source;
+mod stream_rrd_from_http;
 
 #[cfg(not(target_arch = "wasm32"))]
 mod load_stdin;
