+++ crates/store/re_log_encoding/Cargo.toml
@@ -45,7 +45,6 @@ stream_from_http = [
 re_arrow_util.workspace = true
 re_build_info.workspace = true
 re_chunk.workspace = true
-re_log_channel.workspace = true
 re_log_types.workspace = true
 re_log.workspace = true
 re_protos.workspace = true
